+++ _templates/header.php
@@ -1,3 +1,12 @@
+<?php 
+ $page = basename($_SERVER['SCRIPT_NAME']);
+ $pageNotLogged = array('setup.php','index.php','histo.php');
+ 
+ if(!in_array($page,$pageNotLogged) && !session::getInstance()->getVar('logged')){
+    header('Location: /errors/401.php');
+    exit();
+ }
+?>
 <!DOCTYPE html>
 <html lang="fr">
   <head>
@@ -6,7 +15,9 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
 
     <title>OkoVision</title>
-    
+    <script type="text/javascript">
+            var sessionToken = "<?php echo session::getInstance()->getVar('sid'); ?>";		
+   </script>
 	<!-- Bootstrap core CSS -->
     <link href="css/bootstrap.min.css" rel="stylesheet">
     <!-- Bootstrap theme -->
@@ -20,6 +31,7 @@
       <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>+
     <![endif]-->
 	<?php //include_once("analyticstracking.php"); ?>
+	
 	</head>
 
   <body role="document">