+++ _templates/menu.php
@@ -5,16 +5,17 @@
 * Utilisation commerciale interdite sans mon accord
 ******************************************************/
 
-  $page = basename($_SERVER['SCRIPT_NAME']);
+ 
  
  function getMenu(){
-	global $page;
+	//global $page;
+	$page = basename($_SERVER['SCRIPT_NAME']);
 	
 	$menu = array(  'index.php' => array(
-	                                    'txt' => session::getLabel('lang.text.menu.index'),
+	                                    'txt' => session::getInstance()->getLabel('lang.text.menu.index'),
 	                                    'icon' => 'glyphicon glyphicon-dashboard'),
 					'histo.php' => array(
-					                    'txt' => session::getLabel('lang.text.menu.historic'),
+					                    'txt' => session::getInstance()->getLabel('lang.text.menu.historic'),
 					                    'icon' => 'glyphicon glyphicon-stats')
 			);	
 	
@@ -25,7 +26,7 @@
 	}
 }
 
- ?>
+?>
  <!-- Fixed navbar -->
     <div class="navbar navbar-inverse navbar-fixed-top" role="navigation">
       <div class="container">
@@ -39,38 +40,50 @@
           <a class="navbar-brand" href="#">OkoVision</a>
         </div>
         <div class="navbar-collapse collapse">
-          <ul class="nav navbar-nav">
+          <ul class="nav navbar-nav top-nav">
             <?php getmenu(); ?>
 			
 			<li class="dropdown">
-			    <!--a href="#" data-toggle="modal" data-target="#login-modal"> 
+			  <?php if( !session::getInstance()->getVar('logged') ){   ?>
+			    <a href="#" data-toggle="modal" data-target="#login-modal"> 
 			        <span class="glyphicon glyphicon-user" aria-hidden="true"></span>
-			     </a-->
+			     </a>
+			  <?php } else { ?>
                 <a href="#" class="dropdown-toggle" data-toggle="dropdown">
                     <span class="glyphicon glyphicon-cog" aria-hidden="true"></span>
                 </a>
-                <ul class="dropdown-menu" role="menu">
-                    <li class="dropdown-header"><span class="glyphicon glyphicon-edit" aria-hidden="true"></span> <?php echo session::getLabel('lang.text.menu.graphic') ?></li>
-                        <li><a href="gstrapport.php"><?php echo session::getLabel('lang.text.menu.graphic.report') ?></a></li>
+                <ul class="dropdown-menu">
+                    <li class="dropdown-header"><span class="glyphicon glyphicon-edit" aria-hidden="true"></span> <?php echo session::getInstance()->getLabel('lang.text.menu.graphic') ?></li>
+                        <li><a href="gstrapport.php"><?php echo session::getInstance()->getLabel('lang.text.menu.graphic.report') ?></a></li>
                     <li class="divider"></li>
                    
-                        <li class="dropdown-header"><span class="glyphicon glyphicon-download-alt" aria-hidden="true"></span> <?php echo session::getLabel('lang.text.menu.manual') ?></li>    
-                            <?php if(GET_CHAUDIERE_DATA_BY_IP){ ?><li><a href="amImpBoiler.php"><?php echo session::getLabel('lang.text.menu.manual.import.ip') ?></a></li> <?php } ?>
-                            <li><a href="amImpUsb.php"><?php echo session::getLabel('lang.text.menu.manual.import.usb') ?></a></li>
-                            <li><a href="amImportMass.php"><?php echo session::getLabel('lang.text.menu.manual.import.mass') ?></a></li>
-                            <li><a href="amSynthese.php"><?php echo session::getLabel('lang.text.menu.manual.synthese') ?></a></li>
+                        <li class="dropdown-header"><span class="glyphicon glyphicon-download-alt" aria-hidden="true"></span> <?php echo session::getInstance()->getLabel('lang.text.menu.manual') ?></li>    
+                            <?php if(GET_CHAUDIERE_DATA_BY_IP){ ?><li><a href="amImpBoiler.php"><?php echo session::getInstance()->getLabel('lang.text.menu.manual.import.ip') ?></a></li> <?php } ?>
+                            <li><a href="amImpUsb.php"><?php echo session::getInstance()->getLabel('lang.text.menu.manual.import.usb') ?></a></li>
+                            <li><a href="amImportMass.php"><?php echo session::getInstance()->getLabel('lang.text.menu.manual.import.mass') ?></a></li>
+                            <li><a href="amSynthese.php"><?php echo session::getInstance()->getLabel('lang.text.menu.manual.synthese') ?></a></li>
                     <li class="divider"></li>
                     
-                        <li class="dropdown-header"><span class="glyphicon glyphicon-wrench" aria-hidden="true"></span> <?php echo session::getLabel('lang.text.menu.admin') ?></li>
-                            <li><a href="adminParam.php"><?php echo session::getLabel('lang.text.menu.admin.information') ?></a></li>
-                            <li><a href="adminSeason.php"><?php echo session::getLabel('lang.text.menu.admin.season') ?></a></li>
-                            <li><a href="adminMatrix.php"><?php echo session::getLabel('lang.text.menu.admin.matrix') ?></a></li>
+                        <li class="dropdown-header"><span class="glyphicon glyphicon-wrench" aria-hidden="true"></span> <?php echo session::getInstance()->getLabel('lang.text.menu.admin') ?></li>
+                            <li><a href="adminParam.php"><?php echo session::getInstance()->getLabel('lang.text.menu.admin.information') ?></a></li>
+                            <li><a href="adminSeason.php"><?php echo session::getInstance()->getLabel('lang.text.menu.admin.season') ?></a></li>
+                            <li><a href="adminMatrix.php"><?php echo session::getInstance()->getLabel('lang.text.menu.admin.matrix') ?></a></li>
                     <li class="divider"></li>
                     
-                        <li><a href="about.php"><span class="glyphicon glyphicon-info-sign" aria-hidden="true"></span> <?php echo session::getLabel('lang.text.menu.about') ?></a></li>
+                        <li><a href="about.php"><span class="glyphicon glyphicon-info-sign" aria-hidden="true"></span> <?php echo session::getInstance()->getLabel('lang.text.menu.about') ?></a></li>
                    
                     
                 </ul>
+              </li>
+              <li class="dropdown">
+                <a href="#" class="dropdown-toggle" data-toggle="dropdown">
+			        <span class="glyphicon glyphicon-user" aria-hidden="true"></span>
+			    </a>
+			    <ul class="dropdown-menu">
+			        <li><a href="#" data-toggle="modal" data-target="#password-modal"><?php echo session::getInstance()->getLabel('lang.text.menu.password') ?></a></li>
+			        <li><a id="btlogout" href=""><?php echo session::getInstance()->getLabel('lang.text.menu.logout') ?></a></li>
+			    </ul>    
+            <?php } ?>    
             </li>
 			
            </ul>
@@ -97,20 +110,42 @@
     	  <div class="modal-dialog">
     	      <div class="modal-content">    
         	    <div class="modal-header">
-    			    <h2 >Espace membre</h2>
+    			    <h2>Espace membre</h2>
                 </div>
                 <div class="modal-body">
-                    <form class="form-signin">
-                        <p><label for="inputEmail" class="sr-only">Identifiant</label>
-                        <input type="email" id="inputEmail" class="form-control" placeholder="Identifiant" required autofocus>
+                    <form id="formlogin" class="form-signin">
+                        <p><label for="inputUser" class="sr-only">Identifiant</label>
+                        <input type="text" id="inputUser" class="form-control" placeholder="Identifiant" required autofocus>
                         <label for="inputPassword" class="sr-only">Mot de passe</label>
                         <input type="password" id="inputPassword" class="form-control" placeholder="Mot de passe" required></p>
-                        <p><button class="btn btn-lg btn-primary btn-block" type="submit">Login</button></p>
-                        <brr:>
+                        <p><button class="btn btn-lg btn-primary btn-block"id="btlogin">Login</button></p>
+                        <br/>
                     </form>
+                    
     			</div>
     	    </div>  	
 		  </div>
+    </div>
+    
+    <div class="modal fade" id="password-modal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true" >
+	  <div class="modal-dialog">
+	      <div class="modal-content">    
+    	    <div class="modal-header">
+			    <h2>Changer mon mot de passe</h2>
+            </div>
+            <div class="modal-body">
+                <form id="formChangePass" class="form">
+                    <p><label for="inputPass" class="sr-only">Mot de passe</label>
+                    <input type="password" id="inputPass" class="form-control" placeholder="Mot de passe" required autofocus>
+                    <label for="inputPassConfirme" class="sr-only">Confirmation</label>
+                    <input type="password" id="inputPassConfirm" class="form-control" placeholder="Confirmation" required></p>
+                    <p><button class="btn btn-lg btn-primary btn-block"id="btChangePass">Changer</button></p>
+                    <br/>
+                </form>
+                
+			</div>
+	    </div>  	
+	  </div>
     </div>
     
     <br/>