+++ simu_upgrade.php
@@ -5,6 +5,7 @@ class simu_upgrade extends connectDb{
     
     public function __construct() {
 		parent::__construct();
+		$version = "1.6.0";
 		require('_upgrade.php');
 	}